优化保养计划和保养工单功能，待完善

+++ aspnet-core/DeviceManagementSystem.Utils/Common/PageRequest.cs
@@ -5,12 +5,12 @@
         /// <summary>
         /// 当前页码
         /// </summary>
-        public long Current { get; set; } = 1;
+        public int Current { get; set; } = 1;
 
         /// <summary>
         /// 每页大小
         /// </summary>
-        public long Size { get; set; } = 10;
+        public int Size { get; set; } = 10;
 
         /// <summary>
         /// 排序字段
